Rename EquivalenceTests to FunctionalEquivalenceTests, add functional equivalence test for nullable value types
